Add hydra-jetty for use in development and testing.
Our actual strategy will be to use the hydra-jetty submodule directory
as a template which we'll clone (via rake tasks) to provide development
and testing versions (with different ports). I want to do it this way
in order to facilitate testing - just throw away the old one and put in
a fresh copy each time to assure that we run the tests without any
conflicts from old runs. Of course this will take a little while, but
seems the simplest solution.