Use STI on FileGroup with three subclasses for external, bit-level, and object-level file groups. This also allowed removal of the storage_level attribute, which is implicit from the class
This required:
--
fixing up path helpers in the views, etc. (Since we still
only use the the FileGroupController we need to route subclasses through
to that as well and not just use the default. So we need to replace things
like link_to 'File Group', @file_group with explicit file_group_path(@file_group)

fixing up some stuff with assessments (in the same vein as the above)

some other small changes
--

This will allow us to refactor some code into the subclasses when it is only
applicable there, but that isn't done quite yet.

STI and nested attributes didn't work out so well, so we explicitly update
the collection and rights information when creating/updating a file group.